pyproject.toml
[project]
name = "jlab-mcp"
version = "0.1.0"
description = "Jupyter Lab MCP server for SLURM clusters"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.10.6",
    "nbformat>=5.10.4",
    "requests>=2.32.4",
    "pillow>=11.3.0",
    "psutil>=7.0.0",
    "websocket-client>=1.7.0",
    "jupyterlab>=4.0",
    "ipykernel>=6.30.0",
    "matplotlib>=3.8",
    "numpy>=1.26",
]

[project.scripts]
jlab-mcp = "jlab_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-timeout>=2.2",
]
